requirements.txt
attrs~=20.3.0
allure-pytest~=2.8.40
allure-python-commons~=2.8.40
atomicwrites~=1.4.0
colorama~=0.4.4
chardet~=4.0.0
certifi~=2020.12.5
chardet~=4.0.0
idna~=2.10
iniconfig~=1.1.1
pip~=21.1
py~=1.10.0
pytest~=6.2.3
pytest-html~=3.1.1
pytest-metadata~=1.11.0
pluggy~=0.13.1
packaging~=20.9
pyparsing~=2.4.7
requests~=2.25.1
six~=1.15.0
setuptools~=56.0.0
toml~=0.10.2
urllib3~=1.26.4